updated shared delete confirmation propmpts + settings
