+++ src/mongoClusters/commands/dropCollection.ts
@@ -4,8 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { type IActionContext } from '@microsoft/vscode-azext-utils';
-import { getConfirmationWithWordQuestion } from '../../utils/dialogsConfirmations';
+import { getConfirmationAsInSettings } from '../../utils/confirmations';
 import { type CollectionItem } from '../tree/CollectionItem';
+
 
 export async function dropCollection(context: IActionContext, node?: CollectionItem): Promise<void> {
     // node ??= ... pick a node if not provided
@@ -13,10 +14,9 @@ export async function dropCollection(context: IActionContext, node?: CollectionI
         throw new Error('No collection selected.');
     }
 
-    const confirmed = await getConfirmationWithWordQuestion(
+    const confirmed = await getConfirmationAsInSettings(
         `Drop "${node?.collectionInfo.name}"?`,
-        `Drop collection "${node?.collectionInfo.name}" and its contents?\nThis can't be undone.\n\n` +
-            'Please type the name of the collection to confirm:',
+        `Drop collection "${node?.collectionInfo.name}" and its contents?\nThis can't be undone.\n\n`,
         node?.collectionInfo.name,
     );
