+++ src/mongoClusters/commands/dropDatabase.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { type IActionContext } from '@microsoft/vscode-azext-utils';
-import { getConfirmationWithWordQuestion } from '../../utils/dialogsConfirmations';
+import { getConfirmationAsInSettings } from '../../utils/confirmations';
 import { type DatabaseItem } from '../tree/DatabaseItem';
 
 export async function dropDatabase(context: IActionContext, node?: DatabaseItem): Promise<void> {
@@ -13,10 +13,9 @@ export async function dropDatabase(context: IActionContext, node?: DatabaseItem)
         throw new Error('No database selected.');
     }
 
-    const confirmed = await getConfirmationWithWordQuestion(
+    const confirmed = await getConfirmationAsInSettings(
         `Drop "${node?.databaseInfo.name}"?`,
-        `Drop database "${node?.databaseInfo.name}" and its contents?\nThis can't be undone.\n\n` +
-            'Please type the name of the database to confirm:',
+        `Drop database "${node?.databaseInfo.name}" and its contents?\nThis can't be undone.`,
         node?.databaseInfo.name,
     );
